core/authentication/urls.py
from rest_framework_simplejwt import views as jwt_views
from django.urls import path, include
from authentication.api.routers import router


urlpatterns = [
    path("", include(router.urls)),
    path("token/", jwt_views.TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", jwt_views.TokenRefreshView.as_view(), name="token_refresh"),
]
