core/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-smtp-ssl==1.0
python-dotenv==1.0.1